requirements.txt
torch
torchvision
torchmetrics
numpy
pillow
streamlit
requests
opencv-python-headless
streamlit-image-comparison